backend/pyproject.toml
[project]
name = "ai-identity-tracker-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "PyJWT>=2.8.0",
  "SQLAlchemy>=2.0.0",
  "alembic>=1.13.0",
  "python-dotenv>=1.0.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"

[tool.uvicorn]
factory = true


